src/982A/test_cdf_982A.py
import unittest
from unittest.mock import patch

from cdf_982A import CodeforcesTask982ASolution


class TestCDF982A(unittest.TestCase):
    def test_982A_acceptance_1(self):
        mock_input = ['3', '101']
        expected = 'Yes'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask982ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEquals(expected, actual)

    def test_982A_acceptance_2(self):
        mock_input = ['4', '1011']
        expected = 'No'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask982ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEquals(expected, actual)

    def test_982A_acceptance_3(self):
        mock_input = ['5', '10001']
        expected = 'No'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask982ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEquals(expected, actual)


if __name__ == "__main__":
    unittest.main()
